requirements.txt
# Adding python version here blocks the pip install operation
# python_version == 3.11

# FastAPI framework
fastapi~=0.112.2

# ASGI server for FastAPI
uvicorn[standard]~=0.32.1

# Web scraping dependencies
selenium==4.18.1
webdriver-manager==4.0.1  # Automatically handles ChromeDriver updates
requests~=2.32.3

# Data handling and parsing (Optional but useful for scraping)
beautifulsoup4==4.12.3
lxml==5.1.0

# Environment variables (Optional for better config management)
python-dotenv==1.0.1

# Optional logging and monitoring
loguru==0.7.2

pytest~=7.4.4
h11~=0.14.0
jupyterlab-pygments~=0.3.0
pip~=24.2
attrs~=24.2.0
# zlib~=1.2.13
distro~=1.9.0
wheel~=0.44.0
pysocks~=1.7.1
protobuf~=4.25.5
# openssl~=3.3.2
cryptography~=43.0.0
tornado~=6.4.1
jinja2~=3.1.4
typing_extensions~=4.12.2
sniffio~=1.3.1
trio~=0.26.2
outcome~=1.1.0
starlette~=0.38.2
pydantic~=2.9.2
anyio~=4.7.0
orjson~=3.10.11
click~=8.1.7
wsproto~=1.2.0
hypothesis~=6.116.0
deprecated~=1.2.14
annotated-types~=0.7.0
itsdangerous~=2.2.0
pyyaml~=6.0.2
httpx~=0.28.1
sqlalchemy~=2.0.36
pymysql~=1.1.1
apscheduler~=3.11.0
rapidfuzz~=3.12.1
openai~=1.64.0